fix: separate Firework with Particle Component on useFirework.js
