+++ src/components/firework/index.scss
@@ -5,6 +5,6 @@
   display: block;
   width: 100%;
   height: 100%;
-  z-index: 1;
+  z-index: 2;
   background-color: transparent;
 }
